# Artemis: feat: Enhance FIX scanner performance with SIMD, memory, and compiler optimizations