--- a/src/scanner.c
+++ b/src/scanner.c
@@ -32,20 +32,45 @@ EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 #ifdef USE_SSE
 #include <xmmintrin.h>
+#include <emmintrin.h>  // For SSE2 intrinsics
+#include <smmintrin.h>  // For SSE4.1 intrinsics (if available)
 #endif
 
-// message buffer handling
+// Enable prefetch hints for memory access patterns
+#ifdef __GNUC__
+#define PREFETCH(addr) __builtin_prefetch(addr)
+#else
+#define PREFETCH(addr)
+#endif
+
+// Compiler optimization hints
+#ifdef __GNUC__
+#define LIKELY(x)   __builtin_expect(!!(x), 1)
+#define UNLIKELY(x) __builtin_expect(!!(x), 0)
+#else
+#define LIKELY(x)   (x)
+#define UNLIKELY(x) (x)
+#endif
+
+// Optimized buffer growth strategy - grow by 1.5x, minimizes reallocs
 static
 char* make_space(fix_parser* const parser, char* dest, unsigned extra_len)
 {
 	const unsigned n = dest - parser->body, len = n + extra_len;
 
-	if(len > parser->body_capacity)
+	if(UNLIKELY(len > parser->body_capacity))
 	{
+		// Calculate new capacity - grow by 1.5x for better performance
+		unsigned new_capacity = parser->body_capacity * 3 / 2;
+		if (new_capacity < len) new_capacity = len;
+        
+		// Align to cache line size for better memory access patterns
+		new_capacity = (new_capacity + 63) & ~63;
+        
 		// reallocate memory
-		char* const p = realloc(parser->body, len);
+		char* const p = realloc(parser->body, new_capacity);
 
-		if(!p)
+		if(UNLIKELY(!p))
 		{
 			set_fatal_error(parser, FE_OUT_OF_MEMORY);
 			return NULL;
@@ -59,7 +84,7 @@ char* make_space(fix_parser* const parser, char* dest, unsigned extra_len)
 
 		// store new pointer
 		parser->body = p;
-		parser->body_capacity = len;
+		parser->body_capacity = new_capacity;
 	}
 
 	return dest;
@@ -68,18 +93,26 @@ char* make_space(fix_parser* const parser, char* dest, unsigned extra_len)
 // initialisation
 bool init_scanner(fix_parser* parser)
 {
-	char* const p = malloc(INITIAL_BODY_SIZE);
+	// Align initial buffer to cache line size
+	const unsigned aligned_size = (INITIAL_BODY_SIZE + 63) & ~63;
+	
+	#ifdef __GNUC__
+	// Use aligned_alloc for better memory alignment if available
+	char* const p = aligned_alloc(64, aligned_size);
+	#else
+	char* const p = malloc(aligned_size);
+	#endif
 
-	if(!p)
+	if(UNLIKELY(!p))
 		return false;
 
 	parser->body = p;
-	parser->body_capacity = INITIAL_BODY_SIZE;
+	parser->body_capacity = aligned_size;
 	return true;
 }
 
-// scanner helper functions
-static inline
+// scanner helper functions - force inlining for critical path functions
+static inline __attribute__((always_inline))
 unsigned min(unsigned a, unsigned b)
 {
 	return a < b ? a : b;
@@ -107,39 +140,61 @@ unsigned char copy_cs(char* restrict p, const char* restrict s, unsigned n)
 #ifdef USE_SSE
 	if(end - s >= 16)
 	{
-		__m128i cs128 = _mm_loadu_si128((const __m128i*)s);
+		PREFETCH(s + 64);  // Prefetch ahead for next iteration
+		__m128i cs128 = _mm_setzero_si128();  // Initialize checksum to zero
 
-		_mm_storeu_si128((__m128i*)p, cs128);
-
-		for(s += 16, p += 16; end - s >= 16; s += 16, p += 16)
+		// Process 16 bytes at a time
+		for(; end - s >= 16; s += 16, p += 16)
 		{
 			const __m128i tmp = _mm_loadu_si128((const __m128i*)s);
-
 			_mm_storeu_si128((__m128i*)p, tmp);
 			cs128 = _mm_add_epi8(cs128, tmp);
 		}
 
+		// Horizontal sum
+		#ifdef __SSE4_1__
+		// More efficient horizontal sum with SSE4.1
 		cs128 = _mm_add_epi8(cs128, _mm_srli_si128(cs128, 8));
 		cs128 = _mm_add_epi8(cs128, _mm_srli_si128(cs128, 4));
 		cs128 = _mm_add_epi8(cs128, _mm_srli_si128(cs128, 2));
 		cs128 = _mm_add_epi8(cs128, _mm_srli_si128(cs128, 1));
-		cs += _mm_extract_epi16(cs128, 0);	// SSE4: _mm_extract_epi8 ?
+		cs += _mm_extract_epi8(cs128, 0);
+		#else
+		// SSE2 fallback
+		cs128 = _mm_add_epi8(cs128, _mm_srli_si128(cs128, 8));
+		cs128 = _mm_add_epi8(cs128, _mm_srli_si128(cs128, 4));
+		cs128 = _mm_add_epi8(cs128, _mm_srli_si128(cs128, 2));
+		cs128 = _mm_add_epi8(cs128, _mm_srli_si128(cs128, 1));
+		cs += _mm_extract_epi16(cs128, 0) & 0xFF;
+		#endif
 	}
 
+	// Process 8 bytes at a time for remaining data
 	if(end - s >= 8)
 	{
 		__m128i cs64 = _mm_loadl_epi64((const __m128i*)s);
-
 		_mm_storel_epi64((__m128i*)p, cs64);
+		
 		s += 8;
 		p += 8;
+		
+		#ifdef __SSE4_1__
+		// More efficient horizontal sum with SSE4.1
 		cs64 = _mm_add_epi8(cs64, _mm_srli_si128(cs64, 4));
 		cs64 = _mm_add_epi8(cs64, _mm_srli_si128(cs64, 2));
 		cs64 = _mm_add_epi8(cs64, _mm_srli_si128(cs64, 1));
-		cs += _mm_extract_epi16(cs64, 0);	// SSE4: _mm_extract_epi8 ?
+		cs += _mm_extract_epi8(cs64, 0);
+		#else
+		// SSE2 fallback
+		cs64 = _mm_add_epi8(cs64, _mm_srli_si128(cs64, 4));
+		cs64 = _mm_add_epi8(cs64, _mm_srli_si128(cs64, 2));
+		cs64 = _mm_add_epi8(cs64, _mm_srli_si128(cs64, 1));
+		cs += _mm_extract_epi16(cs64, 0) & 0xFF;
+		#endif
 	}
 #endif	// #ifdef USE_SSE
 
+	// Handle remaining bytes
 	while(s < end)
 		cs += (*p++ = *s++);
 
@@ -158,17 +213,17 @@ bool copy_chunk_cs(scanner_state* const state)
 	return state->counter == 0;
 }
 
-// covert and validate message length
+// covert and validate message length - optimized for branch prediction
 static
 bool convert_message_length(scanner_state* const state)
 {
-	if(state->counter < 2)
+	if(UNLIKELY(state->counter < 2))
 		return false;
 
 	const char* s = state->dest - state->counter;
 	unsigned len = CHAR_TO_INT(*s++) - '0';
 
-	if(len > 9)
+	if(UNLIKELY(len > 9))
 		return false;
 
 	const char* const end = state->dest - 1;
@@ -177,16 +232,16 @@ bool convert_message_length(scanner_state* const state)
 	{
 		const unsigned t = CHAR_TO_INT(*s++) - '0';
 
-		if(t > 9)
+		if(UNLIKELY(t > 9))
 			return false;
 
 		len = len * 10 + t;
 
-		if(len > MAX_MESSAGE_LENGTH)
+		if(UNLIKELY(len > MAX_MESSAGE_LENGTH))
 			return false;
 	}
 
-	if(len < sizeof("35=0|49=X|56=Y|34=1|") - 1)
+	if(UNLIKELY(len < sizeof("35=0|49=X|56=Y|34=1|") - 1))
 		return false;
 
 	state->counter = len;
@@ -227,7 +282,7 @@ bool extract_next_message(fix_parser* const parser)
 				return (state->label = 1, false);
 
 			// validate header
-			if(memcmp(parser->header, parser->body, parser->header_len) != 0)
+			if(UNLIKELY(memcmp(parser->header, parser->body, parser->header_len) != 0))
 				goto BEGIN_STRING_FAILURE;
 
 			state->check_sum = parser->header_checksum;
@@ -241,7 +296,7 @@ bool extract_next_message(fix_parser* const parser)
 			// copy bytes
 			for(;;)
 			{
-				if(state->src == state->end)
+				if(UNLIKELY(state->src == state->end))
 					return (state->label = 2, false);
 
 				const unsigned char x = (*state->dest++ = *state->src++);
@@ -252,12 +307,12 @@ bool extract_next_message(fix_parser* const parser)
 				if(x == SOH)
 					break;
 
-				if(state->counter == 10)	// max. 9 digits + SOH
+				if(UNLIKELY(state->counter == 10))	// max. 9 digits + SOH
 					goto MESSAGE_LENGTH_FAILURE;
 			}
 
 			// convert
-			if(!convert_message_length(state))
+			if(UNLIKELY(!convert_message_length(state)))
 				goto MESSAGE_LENGTH_FAILURE;
 
 			// store context
@@ -266,7 +321,7 @@ bool extract_next_message(fix_parser* const parser)
 			// ensure enough space for the message body
 			parser->frame.begin = state->dest = make_space(parser, state->dest, state->counter + sizeof("10=123|") - 1);
 
-			if(!state->dest)
+			if(UNLIKELY(!state->dest))
 				return false;	// out of memory
 			// fall through
 
@@ -276,7 +331,7 @@ bool extract_next_message(fix_parser* const parser)
 				return (state->label = 3, false);
 
 			// validate
-			if(*(state->dest - 1) != SOH)
+			if(UNLIKELY(*(state->dest - 1) != SOH))
 			{
 				set_error(&parser->result.error, FE_INVALID_MESSAGE_LENGTH, 9);	// preserving the error context
 				return false;
@@ -298,11 +353,11 @@ bool extract_next_message(fix_parser* const parser)
 			parser->body_length = state->dest - parser->body;
 
 			// validate
-			if(*(const unsigned*)(state->dest - 8) != (SOH | ('1' << 8) | ('0' << 16) | ('=' << 24)) || state->dest[-1] != SOH)
+			if(UNLIKELY(*(const unsigned*)(state->dest - 8) != (SOH | ('1' << 8) | ('0' << 16) | ('=' << 24)) || state->dest[-1] != SOH))
 				goto TRAILER_FAILURE;
 
 			// compare checksum
-			if(!valid_checksum(state))
+			if(UNLIKELY(!valid_checksum(state)))
 			{	// invalid checksum - a recoverable error
 				set_error(&parser->result.error, FE_INVALID_VALUE, 10);
 				parser->result.error.context.end = state->dest - 1;
@@ -339,6 +394,3 @@ EXIT:
 	parser->body_length = state->dest - parser->body;
 	return false;
 }
-
-
-