This PR introduces significant performance enhancements to the FIX scanner by applying a combination of SIMD, memory management, and compiler optimizations.

**Key Changes:**

*   **SIMD (SSE) Integration:** Utilizes SSE2 and SSE4.1 intrinsics to optimize data copying and checksum calculation. Checksum calculation is now integrated directly into the SSE data copy loops using parallel additions. Improved horizontal sum logic and added prefetch hints within the SSE loop to boost cache efficiency.
*   **Memory Management:** Implemented an optimized buffer growth strategy using a 1.5x growth factor (`* 3 / 2`) in `make_space` to reduce `realloc` frequency. Ensures buffer memory is aligned to cache line size (64 bytes) using `aligned_alloc` where available, improving data access performance for SIMD operations.
*   **Compiler Optimizations:** Defined and applied `PREFETCH`, `LIKELY`, and `UNLIKELY` macros using compiler builtins. These are used to guide branch prediction and speculative prefetching, particularly in critical or unlikely code paths. Forced inlining of the `min` helper function to eliminate call overhead.

These optimizations collectively aim to reduce memory allocation costs, speed up core data processing loops, and improve overall code execution flow, leading to faster and more efficient FIX message parsing.



<details>
<summary>Detailed Score Information</summary>

### Score Details

This section contains detailed information about the performance scores for top 5 scored suggestions.

## Top Performing Changes

### 1. src/scanner.c:1-344 - Mean Improvement: 0.44, Mean Original Score: 4.11

- 🟢 **Performance** (Score: 4.70; Change: +1.70): The changes introduce several significant performance optimizations: 1) Improved memory management with a smarter buffer growth strategy (1.5x growth instead of exact sizing), 2) Memory alignment to cache lines (64 bytes) for better CPU cache utilization, 3) Memory prefetching for improved cache hit rates, 4) Branch prediction hints (LIKELY/UNLIKELY macros) to optimize CPU pipeline, 5) Enhanced SIMD optimizations including proper SSE4.1 path with _mm_extract_epi8 when available, 6) Function inlining attributes for critical path functions, and 7) Better horizontal sum implementation in SSE code. These changes collectively should reduce memory allocation overhead, improve cache utilization, reduce pipeline stalls, and optimize vector processing, leading to measurably better performance for parsing high-throughput FIX messages.

- 🟢 **[LOG] Build Runtime Efficiency** (Score: 4.26; Change: +0.06): The provided code is optimized for performance with several strategies, such as using SSE intrinsics for parallel processing, prefetching, and minimizing memory reallocations. However, the total runtime of the build process is 1.939 seconds, which indicates there may still be room for improvement in runtime efficiency. Further profiling and optimization could help reduce this time, particularly by analyzing the memory access patterns and the efficiency of the algorithms used.

- 🟢 **[LOG] CPU Utilization Efficiency** (Score: 4.60; Change: +0.40): The code demonstrates excellent CPU efficiency based on the test logs. The runtime was 1.528s with a CPU usage of 66.74%, which indicates the code is efficiently using CPU resources without excessive utilization. The optimizations in the code that contribute to this efficiency include: use of SSE/SIMD instructions for vectorized processing, memory prefetching with PREFETCH macros, branch prediction hints with LIKELY/UNLIKELY macros, cache-aligned memory allocation, optimized buffer growth strategy (1.5x), and efficient inline functions. The tests all pass successfully while processing a high volume of messages (up to 6551K messages per second), showing that the code maintains correctness while being CPU-efficient. The code also handles memory management well, as evidenced by the moderate memory usage (about 36MB) despite processing large volumes of data.

- 🟢 **[LOG] CPU Utilization Ratio** (Score: 3.80; Change: +1.60): The code shows a moderate level of CPU efficiency with a CPU usage of 19.59% during the build process. This indicates that while the code is optimized to some extent (e.g., using SSE intrinsics for parallel processing), there may still be room for improvement in terms of CPU utilization. The use of memory prefetching and inlining critical functions suggests an effort to enhance performance, but further optimizations could be explored to reduce CPU usage even more.

- 🟢 **[LOG] Detailed Logging** (Score: 4.40; Change: +0.20): The provided logs show detailed test information that supports high quality debugging and performance monitoring. Each test clearly shows its name, pass/fail status, and the most impressive aspect is that performance tests include comprehensive metrics: number of messages processed (e.g., '500K messages'), total time taken (e.g., '0.126 s'), throughput in messages per second (e.g., '3958K msg/s'), and per-message processing time (e.g., '0.253 µs/msg'). The logs also group tests by categories (Scanner tests, Parser tests, Utils tests) and provide a summary of results. Additionally, there's overall runtime information at the end with CPU and memory usage statistics. This level of detail would be very valuable for identifying performance bottlenecks, which aligns well with the user's goal of improving performance.

- 🟢 **[LOG] Execution Efficiency** (Score: 4.40; Change: +0.20): The code demonstrates excellent execution performance with impressive message processing rates. The test logs show very high throughput: scanner tests process 6,551,000 messages per second (1000K messages in 0.153s), and various parser tests achieve between 1,773,000 and 3,958,000 messages per second depending on message complexity. These are exceptional performance metrics for FIX message processing. The code uses numerous optimizations including SSE instructions, prefetching, compiler hints, cache-line alignment, and optimized buffer growth strategies. The extremely low per-message processing time (as low as 0.153 µs/msg for scanning) indicates highly efficient implementation. All tests pass successfully, showing the code maintains correctness while achieving this performance. The only reason this doesn't receive a perfect score is that there's always room for marginal performance improvements in high-performance computing.

- 🟢 **[LOG] Memory Usage Efficiency** (Score: 4.10; Change: +0.70): The code shows several optimizations aimed at reducing memory usage, such as using aligned memory allocation and a buffer growth strategy that minimizes reallocations. The use of SSE intrinsics for processing data also indicates an effort to optimize performance and memory access patterns. However, the overall memory usage during the build process was 23,336,960 bytes, which could still be improved further. Therefore, the score reflects a good level of optimization but leaves room for improvement.

- 🟢 **[LOG] Performance Consistency** (Score: 4.80; Change: +0.60): The code demonstrates excellent consistency in message processing times across tests. When examining the unit test logs, we can see the same tests were run twice, and the processing times are remarkably consistent. For example, scanner_timed_test processes 1000K messages in 0.153s (6551K msg/s) in both test runs. Similarly, timed_simple_test (0.126s), timed_group_test (0.282s), timed_simple_full_spec_test (0.158s), and timed_full_spec_group_test (0.266s) show identical times across both test runs. This high level of consistency indicates the code has stable performance characteristics with minimal variance, which is especially important for financial applications where predictable performance is critical. The code's optimizations (SSE instructions, prefetching, aligned memory allocation, and branch prediction hints) are likely contributing to this consistency by reducing performance variability.

- 🟢 **[LOG] Test Redundancy** (Score: 2.20; Change: +0.40): The code has a significant issue with duplicate test results in the logs. Looking at the unit test logs, we can see that the entire test suite is run twice, with identical test results appearing in sequence. Every test case is run and reported twice, showing exactly the same performance metrics (e.g., 'scanner_timed_test: 1000K messages in 0.153 s (6551K msg/s, 0.153 µs/msg)' appears twice). This indicates inefficiency in the testing framework or execution configuration, as running the same tests multiple times without any changes between runs wastes time and resources. In the context of the user's goal to improve performance and reduce memory usage, these duplicate test executions are counterproductive.

- 🟢 **[LOG] Test Variety** (Score: 4.60; Change: +0.80): The test suite demonstrates excellent diversity in test cases. It includes tests for the scanner component (simple tests, multiple invocation, various error cases like invalid header, checksum, and message length), parser functionality (simple parsing, group handling, duplicate tags, full specification handling, binary data tests), and utility functions. Performance testing is also thoroughly integrated with timed tests for different message types and complexity levels (simple messages, group messages, full specification). The tests cover normal operation paths, error conditions, and edge cases. The random messages test adds further coverage. This comprehensive suite ensures the code is tested across a wide range of usage scenarios and conditions, giving high confidence in its correctness and performance characteristics.


</details>